Implemented DB connection for get_ride

+++ src/models.rs
@@ -3,12 +3,12 @@ use rocket::serde::{Deserialize, Serialize};
 
 // Ride Struct
 #[derive(Queryable, Selectable)]
-// #[diesel(table_name = crate::schema::rides)]
+#[diesel(table_name = crate::schema::rides)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[derive(Deserialize, Serialize)]
 #[serde(crate = "rocket::serde")]
 pub struct Ride {
-    pub id: usize,
+    pub id: i32,
     pub title: String,
     pub description: String,
 }
